perf: accumulate block prefix-sum in uint64 (complete the overflow-UB fix)

+++ src/hdr_histogram.c
@@ -801,7 +801,7 @@ int hdr_value_at_percentiles(const struct hdr_histogram *h, const double *percen
         values[i] = count_at_percentile > 1 ? count_at_percentile : 1;
     }
 
-    int64_t total = 0;
+    uint64_t total = 0; /* unsigned: no signed-overflow UB when a hostile block sum is added at once */
     size_t at_pos = 0;
 
     if (HDR_LIKELY(h->normalizing_index_offset == 0))
@@ -818,18 +818,18 @@ int hdr_value_at_percentiles(const struct hdr_histogram *h, const double *percen
         for (; idx + BATCH_SCAN_BLOCK <= len && at_pos < length; idx += BATCH_SCAN_BLOCK)
         {
             /* unsigned block sum: avoid signed-overflow UB on hostile decoded counts (cf. AVX2 reducer) */
-            const int64_t s = (int64_t)(
+            const uint64_t s =
                 (uint64_t)counts[idx]     + (uint64_t)counts[idx + 1] +
                 (uint64_t)counts[idx + 2] + (uint64_t)counts[idx + 3] +
                 (uint64_t)counts[idx + 4] + (uint64_t)counts[idx + 5] +
-                (uint64_t)counts[idx + 6] + (uint64_t)counts[idx + 7]);
-            if (total + s >= values[at_pos])
+                (uint64_t)counts[idx + 6] + (uint64_t)counts[idx + 7];
+            if (total + s >= (uint64_t)values[at_pos])
             {
                 int32_t j;
                 for (j = idx; j < idx + BATCH_SCAN_BLOCK; j++)
                 {
-                    total += counts[j];
-                    while (at_pos < length && total >= values[at_pos])
+                    total += (uint64_t)counts[j];
+                    while (at_pos < length && total >= (uint64_t)values[at_pos])
                     {
                         values[at_pos] = highest_equivalent_value(h, hdr_value_at_index(h, j));
                         at_pos++;
@@ -844,8 +844,8 @@ int hdr_value_at_percentiles(const struct hdr_histogram *h, const double *percen
         /* Tail: fewer than BATCH_SCAN_BLOCK counters remain. */
         for (; idx < len && at_pos < length; idx++)
         {
-            total += counts[idx];
-            while (at_pos < length && total >= values[at_pos])
+            total += (uint64_t)counts[idx];
+            while (at_pos < length && total >= (uint64_t)values[at_pos])
             {
                 values[at_pos] = highest_equivalent_value(h, hdr_value_at_index(h, idx));
                 at_pos++;
@@ -859,8 +859,8 @@ int hdr_value_at_percentiles(const struct hdr_histogram *h, const double *percen
         hdr_iter_init(&iter, h);
         while (hdr_iter_next(&iter) && at_pos < length)
         {
-            total += iter.count;
-            while (at_pos < length && total >= values[at_pos])
+            total += (uint64_t)iter.count;
+            while (at_pos < length && total >= (uint64_t)values[at_pos])
             {
                 values[at_pos] = highest_equivalent_value(h, iter.value);
                 at_pos++;
